# BlockChainAppDev/PetBoarding/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: 宠物寄养系统API
  description: |
    宠物寄养系统的RESTful API文档，包含用户、宠物、寄养服务、订单、评价和通知等功能。
    
    ## 认证
    大部分API需要通过JWT令牌进行认证，请在请求头中添加`Authorization: Bearer {token}`。
    
    ## 错误处理
    所有API响应均使用统一的格式，错误时会返回相应的错误码和错误信息。
  version: 1.0.0
  contact:
    name: 宠物寄养系统团队
    email: support@petboarding.example.com
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.petboarding.example.com/api
    description: 生产环境
...
paths:
...
  /users/register:
    post:
      tags:
        - 用户
      summary: 注册用户
      description: 注册新用户
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/login:
    post:
      tags:
        - 用户
      summary: 用户登录
      description: 用户登录并获取令牌
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      tags:
        - 用户
      summary: 获取用户信息
      description: 获取指定用户的信息
      operationId: getUserInfo
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 用户
      summary: 更新用户信息
      description: 更新指定用户的信息
      operationId: updateUserInfo
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 宠物服务API
...
  /pets/user/{userId}:
    get:
      tags:
        - 宠物
      summary: 获取用户的所有宠物
      description: 获取指定用户的所有宠物
      operationId: getUserPets
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有宠物
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 寄养服务API
...
  /orders/user/{userId}:
    get:
      tags:
        - 订单
      summary: 获取用户的所有订单
      description: 获取指定用户的所有订单
      operationId: getUserOrders
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: 订单状态过滤
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reviews/user/{userId}:
    get:
      tags:
        - 评价
      summary: 获取用户的所有评价
      description: 获取指定用户的所有评价
      operationId: getUserReviews
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有评价
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 通知服务API
  /notifications/user/{userId}:
    get:
      tags:
        - 通知
      summary: 获取用户的所有通知
      description: 获取指定用户的所有通知
      operationId: getUserNotifications
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
        - name: read
          in: query
          description: 是否已读，0-未读，1-已读
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有通知
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/users:
    get:
      tags:
        - 管理员
      summary: 获取所有用户
      description: 获取所有用户
      operationId: getAllUsers
      parameters:
        - name: role
          in: query
          description: 角色过滤
          required: false
          schema:
            type: integer
        - name: status
          in: query
          description: 状态过滤
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取所有用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Invalid parameters"
        data:
          type: object
          nullable: true

    # 用户相关
    UserRegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          example: "johndoe"
        password:
          type: string
          example: "password123"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "1234567890"
        role:
          type: integer
          enum: [0, 1]
          default: 0
          example: 0
    UserLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "johndoe"
        password:
          type: string
          example: "password123"
    UserLoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            user_id:
              type: string
              example: "u123456789"
            username:
              type: string
              example: "johndoe"
            role:
              type: integer
              example: 0
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expires_at:
              type: string
              format: date-time
              example: "2023-06-02T12:00:00Z"
    UserInfoResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            user_id:
              type: string
              example: "u123456789"
            username:
              type: string
              example: "johndoe"
            email:
              type: string
              example: "john@example.com"
            phone:
              type: string
              example: "1234567890"
            role:
              type: integer
              example: 0
            status:
              type: integer
              example: 1
            created_at:
              type: string
              format: date-time
              example: "2023-06-01T12:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-01T12:00:00Z"
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "newemail@example.com"
        phone:
          type: string
          example: "9876543210"
        password:
          type: string
          example: "newpassword123"
    UserListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 1000
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            users:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: string
                    example: "u123456789"
                  username:
                    type: string
                    example: "johndoe"
                  email:
                    type: string
                    example: "john@example.com"
                  phone:
                    type: string
                    example: "1234567890"
                  role:
                    type: integer
                    example: 0
                  status:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-01T12:00:00Z"
...
    PetListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: array
          items:
            type: object
            properties:
              pet_id:
                type: string
                example: "p123456789"
              user_id:
                type: string
                example: "u123456789"
              name:
                type: string
                example: "Buddy"
              type:
                type: string
                example: "Dog"
              breed:
                type: string
                example: "Golden Retriever"
              age:
                type: integer
                example: 3
              gender:
                type: integer
                example: 1
              weight:
                type: number
                format: float
                example: 25.5
              description:
                type: string
                example: "Friendly and energetic"
              created_at:
                type: string
                format: date-time
                example: "2023-06-01T14:00:00Z"
              updated_at:
                type: string
                format: date-time
                example: "2023-06-01T14:00:00Z"

    # 寄养服务相关
...
    OrderListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 15
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            orders:
              type: array
              items:
                type: object
                properties:
                  order_id:
                    type: string
                    example: "o123456789"
                  service_id:
                    type: string
                    example: "s123456789"
                  service_title:
                    type: string
                    example: "Luxury Pet Boarding"
                  pet_id:
                    type: string
                    example: "p123456789"
                  pet_name:
                    type: string
                    example: "Buddy"
                  start_date:
                    type: string
                    format: date
                    example: "2023-07-01"
                  end_date:
                    type: string
                    format: date
                    example: "2023-07-05"
                  total_price:
                    type: number
                    format: float
                    example: 250.00
                  status:
                    type: integer
                    example: 1
                  payment_status:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-01T19:00:00Z"
...
    ReviewListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 30
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            average_rating:
              type: number
              format: float
              example: 4.8
            reviews:
              type: array
              items:
                type: object
                properties:
                  review_id:
                    type: string
                    example: "r123456789"
                  user_id:
                    type: string
                    example: "u123456789"
                  user_name:
                    type: string
                    example: "John Doe"
                  service_id:
                    type: string
                    example: "s123456789"
                  service_title:
                    type: string
                    example: "Luxury Pet Boarding"
                  rating:
                    type: integer
                    example: 5
                  content:
                    type: string
                    example: "Excellent service, my pet was very happy"
                  photos:
                    type: array
                    items:
                      type: string
                    example: ["url1.jpg", "url2.jpg"]
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-01T23:00:00Z"

    # 通知相关
...
    NotificationListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 20
            unread_count:
              type: integer
              example: 5
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            notifications:
              type: array
              items:
                type: object
                properties:
                  notification_id:
                    type: string
                    example: "n123456789"
                  type:
                    type: integer
                    example: 1
                  title:
                    type: string
                    example: "订单状态更新"
                  content:
                    type: string
                    example: "您的订单 #o123456789 已被接受"
                  related_id:
                    type: string
                    example: "o123456789"
                  read:
                    type: integer
                    enum: [0, 1]
                    example: 0
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-02T09:00:00Z"

    # 管理员相关
...